text reviewed

+++ delete.php
@@ -10,17 +10,19 @@
     <main>
       <header>
         <div class="Titre">
-          <p><a id="logo" href="acceuil.php"><span class="hp">Harry Potter</span><br>Magical Creatures</p>
+          <p><a id="logo" href="index.php"><span class="hp">Harry Potter</span><br>Magical Creatures</p>
         </div>
       </header>
       <nav>
       <ul class="Menu">
-        <li><a id="AboutUs" href="acceuil.php">Acceuil</a></li>
-        <li><a id="Table" href="harrypotter.php?#Tableau">Créatures</a></li>
-        <li>Gestion
+        <li><a id="AboutUs" href="index.php">Acceuil</a></li>
+        <li><a id="Table" href="harrypotter.php?#Tableau">Liste</a>
+        </li>
+
+        <li><a id="Contribute">Gestion</a>
           <ul>
             <li><a id="sub" href="create.php">Créer</a><li>
-            <li><a id="sub" href="#">Modifier</a><li>
+            <li><a id="sub" href="modif.php">Modifier</a><li>
             <li><a id="sub" href="delete.php">Supprimer</a><li>
           </ul>
         </li>
@@ -28,11 +30,11 @@
       </nav>
 <table id="Tableau">
         <tr>
+        <th class="table">id
+          </th>
           <th class="table">Nom
           </th>
           <th class="table">Description
-          </th>
-          <th class="table">id
           </th>
           <th class="table">image
           </th>
@@ -42,16 +44,16 @@
 <?php
 $rep_txt = "./txt";
 $rep_img = "./img";
-/*if (isset($_POST['supp'])){
+if (isset($_POST['supp'])){
                 if (file_exists($rep_txt."/".$_POST['supp'].".txt") AND file_exists($rep_img."/".$_POST['supp'].".jpg")){
                   unlink ($rep_txt."/".$_POST['supp'].".txt");
                   unlink ($rep_img."/".$_POST['supp'].".jpg");
-                    echo "votre créatures a bien été supprimé";
+                    echo "la créature a bien été supprimé";
                 }
-}*/
+}
+
 if ($dir_txt = opendir($rep_txt)) {
-    
-    echo "<br>";
+
     while ($filename = readdir($dir_txt)) {
         if ($filename != "." && $filename != "..") {
             $tableau = array();
@@ -59,26 +61,21 @@ if ($dir_txt = opendir($rep_txt)) {
             $file = fopen($path, "r");
             while (!feof($file)) {
                 $line                = fgets($file);
-                $separe              = explode(" : ", $line);
-               // $tableau[$separe[0]] = $separe[1];
+                $separe              = explode("*%", $line);
+
             }
             fclose($file);
+
+            $id = $separe[0];
             echo "<tr>";
-            foreach ($tableau as $key => $value) {
-                echo "<td>" . $value . "</td>";
-            }
-            $chemin =$rep_img."/".$separe[0].".JPG";
-            if (isset($_POST['supp'])){
-              if (file_exists($rep_txt."/".$_POST['supp'].".txt") AND file_exists($rep_img."/".$_POST['supp'].".jpg")){
-                unlink ($rep_txt."/".$_POST['supp'].".txt");
-                unlink ($rep_img."/".$_POST['supp'].".jpg");
-                  echo "votre créatures a bien été supprimé";
+            if ($separe[3] == 1){
+              for ($i=0; $i <3 ; $i++) { 
+                echo "<td><a class='tlink' href='individual.php?creature=$separe[0]#Table'>".$separe[$i]."</a></td>";
               }
-}
-         echo"<form action='delete.php' method='post' name='delete'><td><button type='submit' name='supp' value='$separe[0]'>
-            <img src='del.png' alt='' height='42' width='42'></button></td></form>";
-            
+            echo "<td><img src=" . $rep_img . '/' . $id . ".JPG width='100px' ></td>";
+            echo "<form action='delete.php' method='post' name='delete'><td><button class='button1' type='submit' name='supp' value='$id'><img src='delete.png' alt='' height='42' width='42'></button></td></form>";
             echo "</tr>";
+            }
         }
     }
 }
@@ -90,5 +87,4 @@ if ($dir_txt = opendir($rep_txt)) {
 <p id="Copyright">Ont contribués : <cite>GUILLAUME Anais, FALCETTA Nicolas et QEDIRA Fares</cite> (RAN 1-3)</p>
 <a id="up" href="#logo">Haut de Page</a>
 </footer>
-
 </html>
